+++ src/test/java/net/emteeware/common/DigitSumNegativeTest.java
@@ -4,12 +4,12 @@ import org.junit.Test;
 
 /**
  * Created by Merlin Thomas on 2015-11-10.
- *  Copyright 2015 eMTeeWare
+ * ï¿½Copyright 2015 eMTeeWare
  */
 public class DigitSumNegativeTest {
 
     @Test(expected=IllegalArgumentException.class)
     public void testGetDigitSumWithNegativeInput() throws Exception {
-        int result = DigitSum.getDigitSum(-1);
+        long result = DigitSum.getDigitSum(-1);
     }
 }
